feed-aggregator/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
feedparser>=6.0.10
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
pydantic>=2.5.2
schedule>=1.2.1
flask>=3.0.0 